Convert the taskPrototype to a TaskList class (#622)
* Convert the array-like `tasks` object to a TaskList instance

Makes it easier to reason about, because now it is not a mix of an array
and an object.
Instead, the array is kept as a reference inside TaskList,
and the global tasks variable is an instance of it.

Avoids the need of iterating a prototype object and assigning its props
to an array.

* Get rid of the parameter-less TaskList::get(). Provide helpers instead

In some places tasks.get() is used to access a task by its index, e.g.:
tasks.get()[length - 1]. Instead, TaskList should offer a neat way to
accomplish the same but for now, use tasks.tasks[index].

* Implement TaskList::indexOf to simplify search by task

* Simplify the determination of the next/previous task using the modulo op

* Add TaskList::byIndex(), not to access TaskLists internals from outside

* Define getTaskByNameOrNumber in terms of TaskList::find() for simplicity

Removes completely usage of tasks.tasks, i.e. TaskList's internals are
now not accessed from anywhere.

* Simplify TaskList methods by using the provided find() helpers

Converted bare loops to a more functional style for clarity

* Simplify TaskList::update by not throwing away the already found task

Keep a reference to the task found and operate directly on it. No need
to search for it twice.

* Pass a task reference to TaskList::update instead of id

Removes the need for null checks in the update method

* No need to sort tasks to find the latest, reduce() instead

* Remove TaskList::update() as it's not relevant to the TaskList context

Also, it was used only once, with only a single param to update. It's
expressed simpler as done in the commit.

* Fix style of retrieving the nextTask

https://github.com/minbrowser/min/pull/622#discussion_r219024295

* Implement TaskList::find() with a for loop instead of the built in func

https://github.com/minbrowser/min/pull/622#discussion_r218673511

* Fix prev task determination style